refactor: LineItem을 view 패키지로 이동
