+++ src/main/java/util/LineItemGenerator.java
@@ -1,6 +1,6 @@
 package util;
 
-import domain.LineItem;
+import view.LineItem;
 
 public interface LineItemGenerator {
     LineItem generate();
