+++ src/test/java/domain/LineTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import util.TestLineItemGenerator;
+import view.LineItem;
+
 import java.util.List;
 import java.util.stream.Stream;
