+++ src/test/java/util/RandomLineItemGeneratorTest.java
@@ -2,7 +2,7 @@ package util;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.LineItem;
+import view.LineItem;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import java.util.List;
